Accounts joined to reactive data gathering

+++ sea-shell-soap-wiremock/sea-shell-cli-lowers/src/main/java/org/jesperancinha/shell/client/lowers/SeaShellsWSDLLowers.java
@@ -24,6 +24,6 @@ public interface SeaShellsWSDLLowers {
     public Lower lowers(
 
             @WebParam(name = "lowerId", targetNamespace = "")
-                    int lowerId
+                    long lowerId
     );
 }
